requirements.txt
Django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
flake8>=3.9.2,<3.10.0
psycopg2>=2.9.1,<2.10.0
Pillow>=8.3.0,<8.4.0
requests==2.8.1